#778 Moved RotateGizmo into a file of its own.
